Serve HTTP ranges properly (#4)
The ActiveStorage::ProxyController uses the Blob to figure out the byte size of the blob. Sadly, Service does not have a method for `size(key)`, and this is why they elected to load the blob. We can encode the size of the blob into the token instead.

HTTP range is done with lazy Rack bodies, so that the maximum amount of data buffered is 5MB. Again, Service does not have a `stream(key, range)` method either, so we have to improvise - besides, in my experience doing a request to a cloud storage provider is usually more connecting-resolving-credentials overhead (if you are already in the cloud environment, on the same provider) than doing the download itself.