Set the default active-storage service is service_name is not given
Given a few models like

```ruby
class Task < ApplicationRecord
  has_many_attached :files, service: :local
end
```

```ruby
class User < ApplicationRecord
  has_many_attached :files
end
```

and a configuration of `config.active_storage.service = :encrypted_local_disk` inside `config/environments/<environment>.rb`

this will allow a single model to provide a custom service: or use the "default" set in the rails configuration.

This fixes the issue that when a model with attached :files does not use a service explicitly it will default to the rails "global" service.